+++ src/main/java/gaji/service/global/exception/code/BaseErrorCodeDto.java
@@ -1,4 +1,4 @@
-package gaji.service.global.exception;
+package gaji.service.global.exception.code;
 
 import lombok.Builder;
 import lombok.Getter;
